# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cumments API
  version: 0.25.4
  description: |
    Backend-only comment API backed by Matrix. Writes are asynchronous:
    POST/PATCH/DELETE accept a submission and return `202 { "submission_id" }`;
    the projected comment later carries the same `submission_id`.

    Comment reads use the RFC 10008 `QUERY` method. OpenAPI does not support
    custom verbs, so each QUERY operation is represented with the
    `x-cumments-query` extension and is sent as `QUERY` on the wire.
servers:
  - url: http://localhost:7931
...
security: []
paths:
...
  /api/v1/sites/{site_id}/pages/{page_slug}/room:
    parameters:
      - $ref: "#/components/parameters/SiteId"
      - $ref: "#/components/parameters/PageSlug"
    get:
      tags: [Public]
      summary: Comment room metadata and system messages
      operationId: getRoomInfo
      responses:
        "200":
          description: Room info
          content:
            application/json:
              schema:
                type: object
                required: [room_id, member_count, system_messages]
                properties:
                  room_id:
                    type: string
                  name:
                    type: string
                    nullable: true
                  topic:
                    type: string
                    nullable: true
                  avatar_url:
                    type: string
                    nullable: true
                  avatar_thumbnail_url:
                    type: string
                    nullable: true
                    description: >
                      Same image through the 96×96 crop thumbnail variant
                      (signed proxy URL when the proxy is enabled).
                  member_count:
                    type: integer
                  system_messages:
                    type: array
                    items:
                      type: object
                      properties:
                        event_id:
                          type: string
                        event_type:
                          type: string
                        state_key:
                          type: string
                        sender:
                          type: string
                        origin_server_ts:
                          type: integer
                        content_json:
                          type: object
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    operatorToken:
      type: http
      scheme: bearer
    claimToken:
      type: apiKey
      in: header
      name: X-Cumments-Claim-Token
  parameters:
    SiteId:
      name: site_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9-]{1,64}$"
    PageSlug:
      name: page_slug
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9-]{1,64}$"
...
  responses:
    Error:
      description: RFC 9457 problem details response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type:
          type: string
          format: uri-reference
          description: Canonical URI of the problem type
        title:
          type: string
          description: Stable short human-readable name of the type
        status:
          type: integer
          format: int32
          description: HTTP status code, identical to the response status
        detail:
          type: string
          description: Occurrence-specific explanation for the client
        code:
          type: string
          description: Stable machine-readable slug; the fragment (after `#`) of `type`
        details:
          type: object
          nullable: true
...